+++ docs/CHANGE_LOG.md
@@ -4,6 +4,7 @@
 - Added a border around the code container (uses root variables for easy changing).
 - Increased the padding slightly for the "Copy" button.
 - Added a background color the code syntax container (next to the numbers) and added a border radius.
+- Added new public function "addLanguage()", which will add a new language to the library and will auto-render DOM elements found for that language.
 - 
 
 <br>